requirements.txt
flask
peewee
flask-login
flask-bcrypt
flask-wtf
kafka-python
